prevent size increase
