+++ packages/kbn-optimizer/limits.yml
@@ -72,7 +72,7 @@ pageLoadAssetSize:
   expressionTagcloud: 14009
   expressionXY: 45000
   features: 4145
-  feedback: 8100
+  feedback: 6248
   fieldFormats: 64634
   fieldsMetadata: 4901
   files: 6037
