+++ x-pack/platform/plugins/private/feedback/public/src/utils/get_app_details.ts
@@ -6,7 +6,32 @@
  */
 
 import type { CoreStart } from '@kbn/core/public';
-import { APPS_WITHOUT_CATEGORY_PREFIX } from '@kbn/feedback-registry';
+
+/**
+ * Apps listed here will not have their category prefix in the satisfaction question.
+ * e.g. Questions for "metrics:inventory" will ask users about "Infrastructure inventory" instead of "Observability - Infrastructure inventory"
+ */
+export const APPS_WITHOUT_CATEGORY_PREFIX = [
+  'ml:singleMetricViewer',
+  'ml:resultExplorer',
+  'ml:analyticsMap',
+  'ml:anomalyExplorer',
+  'apm',
+  'apm:service-groups-list',
+  'apm:services',
+  'apm:traces',
+  'apm:service-map',
+  'apm:dependencies',
+  'apm:settings',
+  'apm:storage-explorer',
+  'apm:tutorial',
+  'metrics',
+  'metrics:inventory',
+  'metrics:hosts',
+  'metrics:metrics-explorer',
+  'metrics:settings',
+  'metrics:assetDetails',
+] as readonly string[];
 
 /**
  * Get current app details from browser URL and nav links.
